it's all on me
Chris, I was wrong. You were never the reason we failed, it was all on me. I understand how you feel. I didn't think it would be that hard to put in the shooting mechanic. I should have asked the teacher. I was the whole reason that Sandstorm was a failure. I'm so sorry I blamed you for our failure. I don't know if you're going to pull the origin and read this, but if you do, I hope you can see it in your heart to forgive me.